refactor: update logic for rendering view all pages depending on season
