+++ src/components/Dashboard.jsx
@@ -2,6 +2,9 @@ import React, { useState, useEffect } from "react";
 import Navbar from "./Navbar";
 import AnimeRow from "./AnimeRow";
 import { useSelector } from "react-redux";
+import getSeason from "../utils/getSeason";
+import getNextSeason from "../utils/getNextSeason";
+import getLastSeason from "../utils/getLastSeason";
 
 export default function Dashboard() {
   const [error, setError] = useState("");
@@ -15,42 +18,9 @@ export default function Dashboard() {
   useEffect(() => {
     const currSeason = getSeason(new Date().getMonth() + 1);
     setCurrentSeason(currSeason);
-    setNextSeason(getNextseason(currSeason));
+    setNextSeason(getNextSeason(currSeason));
     setLastSeason(getLastSeason(currSeason));
   }, []);
-
-  const getSeason = (month) => {
-    if (month >= 3 && month <= 5) {
-      return "Spring";
-    }
-    if (month >= 6 && month <= 8) {
-      return "Summer";
-    }
-    if (month >= 9 && month >= 11) {
-      return "Fall";
-    }
-    return "Winter";
-  };
-
-  const getNextseason = (monthString) => {
-    const nextSeasons = {
-      Spring: "Summer",
-      Fall: "Winter",
-      Winter: "Spring",
-      Summer: "Fall",
-    };
-    return nextSeasons[monthString];
-  };
-
-  const getLastSeason = (monthString) => {
-    const nextSeasons = {
-      Spring: "Winter",
-      Fall: "Summer",
-      Winter: "Fall",
-      Summer: "Spring",
-    };
-    return nextSeasons[monthString];
-  };
 
   return (
     <div className="dashboard-wrapper">
